pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
version = "0.7.0"
name = "fsm-llm"
description = "Finite State Machines for Large Language Models"
readme = "README.md"
authors = [{name = "Nikolas Markou", email = "nikolasmarkou@gmail.com"}]
license = "GPL-3.0-or-later"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = [
    "llm", "fsm", "finite state machine",
    "large language model", "conversation",
    "ai workflows", "agentic workflows"
]

# Core dependencies (minimal)
dependencies = [
    "loguru>=0.7.3",
    "litellm>=1.82.0,<2.0,!=1.82.7,!=1.82.8",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
]

requires-python = ">=3.10"

# Optional extras
[project.optional-dependencies]
workflows = []  # No additional deps; validation uses plain Python DFS
reasoning = []       # No additional deps; included in core install
agents = []          # No additional deps; included in core install
monitor = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "jinja2>=3.1.0",
]
mcp = ["mcp>=1.0.0"]
otel = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
]
a2a = ["httpx>=0.24.0"]
oolong = ["datasets>=3.0.0"]  # M5 slice 7: OOLONG benchmark ingestion
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.4.0",
    "mypy>=1.5.0",
]
all = [
    "fsm-llm[workflows]",
    "fsm-llm[reasoning]",
    "fsm-llm[agents]",
    "fsm-llm[monitor]",
    "fsm-llm[mcp]",
    "fsm-llm[otel]",
    "fsm-llm[a2a]",
    "fsm-llm[oolong]",
]

[project.urls]
"Homepage" = "https://github.com/NikolasMarkou/fsm_llm"
"Bug Tracker" = "https://github.com/NikolasMarkou/fsm_llm/issues"

# Core CLI tools (always available)
# R7 step 18 (D-PLAN-04): the 4 legacy console scripts now delegate to
# the unified `fsm_llm.cli.main` via thin alias functions. The
# legacy main_cli entrypoints in validator.py / visualizer.py /
# fsm_llm_monitor/__main__.py / meta_cli.py remain importable for
# Python consumers but the console-script binaries route through cli/.
# Old fsm-llm script keeps its own __main__:main_cli for back-compat
# (it has --mode flags that predate the unified subparser; both
# surfaces ship in 0.4.x).
[project.scripts]
fsm-llm = "fsm_llm.__main__:main_cli"
fsm-llm-visualize = "fsm_llm.cli.main:visualize_alias"
fsm-llm-validate = "fsm_llm.cli.main:validate_alias"
fsm-llm-monitor = "fsm_llm.cli.main:monitor_alias"
fsm-llm-meta = "fsm_llm.cli.main:meta_alias"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
fsm_llm = ["py.typed"]
"fsm_llm.stdlib.reasoning" = ["py.typed"]
"fsm_llm.stdlib.workflows" = ["py.typed"]
"fsm_llm.stdlib.agents" = ["py.typed"]
fsm_llm_monitor = ["py.typed", "static/*", "templates/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow",
    "integration: marks integration tests",
    "examples: marks example tests",
    "real_llm: marks tests requiring real LLM API",
]
# Curated DeprecationWarning policy.
#
# Default: errors-on-warning is too aggressive for a library that depends on
# litellm + pydantic, both of which surface upstream DeprecationWarnings we do
# not control. Instead we ignore third-party noise explicitly and let our own
# DeprecationWarnings (R13/I5 epoch shims) flow normally so tests can assert
# them via ``pytest.warns(DeprecationWarning)``.
filterwarnings = [
    # Library default — let warnings be observable.
    "default",
    # Pydantic v2 still surfaces a class-based ``Config`` deprecation from
    # some transitive callers; not our problem.
    "ignore::DeprecationWarning:pydantic.*",
    # litellm pulls in httpx / openai SDKs that emit DeprecationWarnings on
    # legacy event-loop helpers under Python 3.12.
    "ignore::DeprecationWarning:litellm.*",
    "ignore::DeprecationWarning:httpx.*",
    # Asyncio's own ``get_event_loop`` deprecation surfaces from pytest-asyncio
    # internals on 3.12 — not actionable from our test code.
    "ignore:There is no current event loop:DeprecationWarning",
]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "F",     # pyflakes
    "W",     # pycodestyle warnings
    "I",     # isort (import sorting)
    "UP",    # pyupgrade (Python version upgrades)
    "B",     # flake8-bugbear
    "RUF",   # ruff-specific rules
]
ignore = [
    "E402",    # module-level import not at top (from __future__)
    "E501",    # line too long (handled by formatter)
    "RUF013",  # implicit Optional — too many signatures to change safely
    "RUF001",  # ambiguous unicode — intentional in test/prompt data
    "RUF022",  # __all__ not sorted — current order is semantically grouped
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B011", "B017"]  # tests may use assert True/False and assertRaises
"examples/*" = ["E501"]
"src/fsm_llm_monitor/*" = ["RUF012"]

[tool.ruff.lint.isort]
known-first-party = ["fsm_llm", "fsm_llm.stdlib", "fsm_llm.stdlib.reasoning", "fsm_llm.stdlib.workflows", "fsm_llm.stdlib.agents", "fsm_llm_monitor"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
no_implicit_optional = true
disallow_untyped_defs = false
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
